+++ leggedsnake/__init__.py
@@ -31,11 +31,13 @@ from .dynamiclinkage import (
     convert_to_dynamic_linkage,
 )
 from .physicsengine import (
-    video,
-    video_debug,
-    all_linkages_video,
     params,
     World,
+)
+from .worldvisualizer import (
+    all_linkages_video,
+    video,
+    video_debug,
     VisualWorld
 )
 from .show_evolution import load_data, show_genetic_optimization