+++ CHANGELOG.md
@@ -24,6 +24,7 @@ genetic optimization.
 - Visuals:
   - ``update`` method of ``VisualWorld`` replaced by ``visual_update``. It clearly separates physics and display time.
   - Frame rate and physics speed are now independent parameters.
+  - Visuals go to a new file ``worldvisualizer.py``.
 - ``startnstop`` argument of ``evolutionary_optimization_builtin`` may now be the name of the file to use (a string).
 - ``pyproject.toml`` updated with the data of ``setup.cfg``. This is now the recommended metadata for the project.